chore: split reducers
